clippy: use variable names directly in format!
I dunno how I feel about this, since it can only be done for variables
and not for more complicated expressions, so the result is inconsistent
syntax. But I'm willing to go along with clippy for now. Later I need
to audit the heavy use of format! macros and such.