LX: for tax purposes, assignment trades happen at market price!
Phew. I finally have the output identical to what it was this morning.
But now we have significantly better type safety and all tax information
is computed "locally" in a way that will be easy to rearrange and that
does not depend on external information in surprising ways.

Cleaned up weird type confusion around "lots" "opens" and "closes"