checkpoint commit; overhaul "interesting order" logging
This creates a new ledgerx::interesting module which can collect
statistics on orders, accumulate these, etc., and rewrite the contract
logging logic to use this. We now log "best bid" and "accumulated good
bid" for any contracts for which these are interesting.

Still todo is the logic that looks at our current available funds and
restricts orders based on that. But it seemed worth committing at this
point because we've changed so much code, and the output is already more
useful than the old output (which would show the best bid and the result
of clearing all bids, and which would filter more clumsily).

We no longer log asks that are worth matching, though some code to
support this remains and will be deleted. We will replace this with more
intelligent logic where we just compute what price we want to ask at,
regardless of the existing state of the book.