remove old price_ref accessor
This accessor was used internally to split apart a BitcoinPrice into a
tuple. There is no reason to do this. It was also used to confirm that
the price reference wasn't too old, but (a) it was impossible to enforce
that users would use this accessor, and (b) this happens so frequently
that we need to handle it on a per-situation basis, rather than having
an accessor that just panics.