rearrange main loop to have multiple message sources
Previously the main loop was waking up on LX messages and then reacting
to everything else that'd happened between messages. This is bad
architecture and has the potential for weird things to happen when LX is
quiet (e.g. overnight). Also it really isn't workable to allow the main
thread to open orders unless we have a mechanism to signal it from other
parts of the code. (Or we could fish the api key into places it doesn't
belong, but that makes me nervous.)

Also cap IV on open orders at 160%, since 200% was showing some pretty
crazy orders.