coinbase: fix "shutdown on 5% movement" emergency shutoff
The previous shutoff logic could basically only detect instantaneous
price movements, and would respond by panicking the price ticker
thread, which would stop the bad price data but leave the tracker
running for several minutes (possibly with a wrong price reference).

Instead, fix the logic to detect rapid but non-instantaneous price
swings, and add an "emergency shutdown" to the main loop, so that we
cancel all orders and shut down in an orderly fashion.