ledgerx: account for balances less incorrectly when sniping bids
When we run through the list of all contracts, we queue up a bunch of
"Create Order" commands to the main message loop. When doing so, we
should pre-emptively assume they succeed and update our balances
accordingly so that we don't make more orders than we're actually able
to execute.

This unfortunately means that if the orders _fail_, because our bot is
not particularly fast, we will wind up with an incorrect balance. A
later commit should update the `Heartbeat` handler to request the
current balances from LX before doing anything else.